refactor: 重构 mainModal 代码结构
