+++ src/content/components/mainModal/index.js
@@ -1,65 +1,17 @@
-import React, { useState, useRef, useEffect } from "react";
-import { Modal, ConfigProvider, Tabs } from "antd";
+import React from "react";
+import { Modal, Tabs } from "antd";
 import Configuration from "../configuration";
 import Chat from "../chat";
-import DragM from "dragm";
-// import { apiReqs } from '@/api'
+import BuildDragTitle from "../dragTitle";
 
-function BuildTitle(props) {
-  const [modalDom, setModalDom] = useState(null);
-
-  const updateTransform = transformStr => {
-    modalDom.style.transform = transformStr;
-  };
-
-  useEffect(() => {
-    console.log('modalDom useEffect');
-    const modalElement = document.getElementsByClassName("ant-modal-wrap")[0];
-    setModalDom(modalElement);
-  }, []);
-
-  return (
-    <ConfigProvider
-      autoInsertSpaceInButton={false}
-      theme={{
-        token: {
-          colorPrimary: "#0064ff",
-          colorPrimaryHover: "#226AFF",
-          colorPrimaryActive: "#0032BE",
-          colorSuccess: "#00be5e",
-          colorWarning: "#ffaa33",
-          colorError: "#ff4433",
-          colorInfo: "#0064ff",
-          colorText: "#16171A",
-          colorTextSecondary: "#2B2C2F",
-          colorTextTertiary: "#696A6D",
-          colorBorder: "#c7c7c7",
-          colorBgMask: "rgba(48, 48, 48, 0.5)",
-          borderRadius: 3,
-          sizeStep: 5,
-          sizeUnit: 5,
-          wireframe: false,
-        },
-      }}
-    >
-
-    <DragM  updateTransform={updateTransform} >
-      <div>{props.title}</div>
-    </DragM>
-    </ConfigProvider>
-  );
-
-  // 弹窗的关闭图标
-  const customIcon = <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={2} stroke="currentColor"> <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" /> </svg>
-}
 
 function ContentModal(props) {
   const { setVisible, visible } = props;
   const setVisibleHandler = () => {
     setVisible(false);
   }
+  const customIcon = <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={2} stroke="currentColor"> <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" /> </svg>
 
-  console.log(props);
   // Tabs
   const items = [
     {
@@ -76,7 +28,7 @@ function ContentModal(props) {
 
   return (
       <Modal
-        title={<BuildTitle visible={visible} title="ONES Sql-helper" />}
+        title={<BuildDragTitle visible={visible} title="ONES Sql-helper" />}
         visible={visible}
         mask={false}
         onCancel={() => {
@@ -86,7 +38,7 @@ function ContentModal(props) {
         width={360}
         className="M-root-modal"
         id="rootModal"
-        // closeIcon={customIcon} 这个也得传进来，很关键
+        closeIcon={customIcon}
         zIndex="99999"
         wrapClassName="M-root-modal-wappers"
       >
